Modification Status on CloudInstance
The following commit bumps the CloudInstance to V4 by adding a
modification status to the interface. The modification-status
defines a way to highlight status changes to the operator of
modifications to a machine whilst it's running. This differs from
status and agent-status, which are generally defined as a way to
know about changes during provisioning.