Refactor where endpoint resides
The following just refactors where endpoint resides, so it's easier
to reason about the code.

Adding remote entities to remote applications

We only use remote entity token in the output, as a missing
macaroon will then check cause the controller to discharge a new
macaroon.

Move remote entities to model

Following ongoing discussions externally to this repo/PR, it was
decided that remote entities should just be placed directly on the
model, that way we should then allow importing and exporting of
remote entities in a much cleaner way without having to traverse
over a remote application.