Remote app omitempty for status field
The following commit ensures that we use omitempty for the status
field on a remote application. It's highly possible that not all
remote applications will have a status field. It event states this
in the documentation for the export migration in the juju/juju.

The tests also ensure that we test for missing status field and
the code correctly parses the remote application, with and without
a status field.