Add application offer uuid
The following ensures that we add the application offer uuid, along with
the application name and description. This will allow us to update the
information of a application offer during a CMR migration. These fields
are required as they're validated during the migration process.

Additionally we want to keep the same offer uuid during a migration
process. This will then allow us to link up all the offer uuids.