Adding offer connection type
The following adds the offer connection to the description, but doesn't
wire up that connection to the model. The following commits should wire
this up correctly.

The one thing I'm unsure about is the wiring up of the RelationID, I
need to work out what this is for, it probably can be omitted so I'll
need to double check this.