Adding offer connection tests to the model
This commit adds the tests for the offer connections, but also fixes any
issues around missing version numbers and also ensures that we test
against the firewall rules version as well on the model.

There does seem to be two types of importing the model, although both
version correctly, but it seems like there are two styles.