Export endpoints map
The addition of endpoints map, allows us to see both the internal and
external representation of the endpoint. This is required when
reconstructing the application offer when peforming a CMR migration.
Essentially we need the internal representation of the endpoint to
ensure that we don't have invalid offers matching.

Unfortunately we don't have the ability to remove the old endpoints
endpoint as we need to ensure we're backwards compatible with v1
application offers. Considering that bundle export uses this as a well
used live feature, it's better to ensure this isn't broken in terms of
compatibility.