Change application offer endpoints type
The following changes the application offer endpoints type. The type now
represents a map[string]string, from a []string. We require both the
internal and external name for the endpoint so we can perform CMR offers
in juju correctly. The issue was the previous code changed this so we
exposed an alternative member to the consumer of the library. This isn't
the intended use of the the description package. Instead the package is
meant to be pure, in that the latest entity interfaces and structs are
the final representation of the serialised type. Having both endpoints
AND endpointsMap removed the purity of the type and caused ambiguity
about which method you actually required or wanted.

The decision should then be that it's on the consumer side of things to
correctly handle the different versioning, even if that means handling
more complexity when doing so!