Update README.md
In an effort to update expectations of what the juju package is intended
to represent, I've updated the README.md to help with that. I hope that
others can then understand that the package is more of a purity package
rather than a way to shim out complexity from the Juju package.