Optional external controller alias
As juju doesn't always talk about controller names directly, mostly as
UUIDs we don't always have an alias for controllers. With that in mind
it is best not to request the alias in a map if it's empty.

The code changes that setup and instead only asks for the alias if it's
found. Otherwise the alias is ignored and a default string valus is
used.

This should prevent the panic during migrating consumers to another
controller.