Add additional fields to charm origin
The additional fields allow us to model a charm from the new charmhub
using the id, revision or channel. With the source of the charm origin
allowing us to understand what to do with the charm with in juju.

The only surpless item in charmorigin is the hash, but I believe this
will be very useful for us modeling the charm payload.