Replace legacy ports with port ranges in machine model
This change bumps the version of the machine seriazliation format and
replaces the legacy OpenedPorts/AddOpenedPorts methods with
OpenedPortRanges variants.

In addition, the V1 machine parsing code will transparently convert V1
ports to V2 port ranges behind the scenes. The conversion logic assumes
that ports are opened for all endpoints. This matches the current
behavior of pre 2.9 juju controllers where open ports are stored in a
document with an empty subnet ID.