CloudInstance virt type
Allow the setting and getting of the virt-type for a cloud instance.
We now require that a virt-type is set in the hardware characteristics
of an instance. This means that for migration import and export we also
need to send that information over the wire.

This PR exposes the virt-type for that to work.