Application lease
Add leases to applications so we can migrate the lease. This will
keep leases the same on applications when we perform model migrations.
This should elevate the need to claim a leader when performing the
application migration for a model and addition will prevent needless
churn of the leaders when migrating.