Introduces agent version as a top level key.
Moving to DQlite in Juju 4.0 we are removing agent version from model
config and making it a first class attribute of a model. To support
migrating from 3.6 and 4.0 we need the agent version to be available at
the root of model migration.

To do this new uses of v11 will have to set the key and for older
versions of the description package an attempt will be made to pull this
information from the model config.