feat: add charm metadata to application description
To correctly have RI (referential integrity) when adding a application
in 4.0, we need to know the charm metadata. This essential metadata
is a requirement for things to be able to function correctly. We
don't need actions, config and lxd profiles, they can be supplied
when we upload the actual binary blob.

As this a deeply nested struct, I'm treating it as one amorphous blob,
as apposed to creating and setting each nested entity as it's own
thing.

The version will be the same for the whole of the metadata in it's
first implementation. Although each nested type has it's own schema,
so future iterations can update it independently from each other.