feat: charm configs
This adds charm configs to the description package. It should be
noted that there already exists a CharmConfig in the application
already, but that is the current values of the config. What this
is representing is the actual charm config. It should be noted
that CharmConfig was originally called settings, which doesn't
help either.