fix(instancedata): remove validation checks for instance data
Migration is currently failing on the PR https://github.com/juju/juju/pull/18214
because the validation (that runs *after* we fill the model on the
legacy state but *before* we fill it with the new domain migrations)
returns an error when trying to migrate a machine that's not
provisioned.

While this check is indeed OK, the fact that we are performing it in a
serialization library is a bad idea. Business logic should only live in
juju, not here.

This patch removes the instance Validation() from machine Validation().