feat: add CIDR support to Address structure
This commit adds a new CIDR field to the Address structure and updates
relevant functions and tests to handle this new field.
The version of the Address structure is incremented to 3 to reflect this change.

CIDR is required as part of address description for
PR https://github.com/juju/juju/pull/18297,linked to
Launchpad bug: https://bugs.launchpad.net/juju/+bug/2073986