refactor: remove model status from description
This commit removes model status and status history from the top level
model in the description package. Status for other entities within a
model such as machine has been left un-touched.

With the move to Juju 4.0 a model's status is now computed for every
request as it is considered display information only. This means because
we can now compute the value we know longer need to store and forward
this when migrating and exporting models from Juju 4.0 onwards.

As part of this removal and change I have also introduced a new version
13 of description to represent the version for which this information
was removed.