feat: resource description v2
This changes adds V2 of the resource description.

Previously, resources had optional information about the "path" and
"description" of the resource. This information is contained in the
charm resource metadata and not needed in resource migration.

The FingerprintHex field name is changed to SHA384 and the Username to
RetrievedBy. This more accuratly reflects the contents of the field

This change also adds kubernetes-application-revisions. These are added
to migrate information about the current resource revision used by a
kubernetes application, as this may differe from the application
revision.

The size, fingerprint and retrieved-by of the resource are not migrated
to the new database using the description package, however, they are
fetched from the model description by the migrationmaster facade to send
to the migrationmaster worker which does the upload of the binary
resource blob via HTTP. It is in these HTTP headers that this
information is transmitted to the new controller. For this reason, it
stays in the description package.

Bundles also make use of this package, however they only touch the
Origin and Revision fields of the resource, which are unaffected by
these changes.